requirements.txt
# Django>=3.0.6,<3.1
Django>=5.0.4,<5.1
djangorestframework
uWSGI
PyPDF2>=3.0.1,<3.1
python-docx>=1.1.0
psycopg2>=2.9.9
requests==2.31.0
celery
django_celery_results
django_celery_beat
redis